+++ src/errors/mod.rs
@@ -1,3 +1,5 @@
+// This module contains all the possible OpenLimits errors.
+
 mod missing_implementation_content;
 mod open_limit_error;
