+++ src/lib.rs
@@ -1,4 +1,4 @@
-//! A open source Rust high performance cryptocurrency trading API with support for multiple exchanges and language wrappers. 
+//! An open source Rust high performance cryptocurrency trading API with support for multiple exchanges and language wrappers.
 //! Focused in safety, correctness and speed.
 //!
 //!## Project goals
@@ -8,7 +8,7 @@
 //! It is initially focused on spot exchanges as a method to define how initial data structures and project architecture 
 //! should be to allow zero cost abstractions 
 //! around the exchanges' peculiarities. So initial project goals are:
-
+//!
 //!* Based on Rust, memory safe by default.
 //!* Support for websockets and user defined networking.
 //!* Thin layer wrappers for Java, C#, Python and Node.js
@@ -16,25 +16,24 @@
 //!* Open-source only, now and forever.
 //!
 //!Future goals are:
-
+//!
 //!* Support for futures trading
 //!* Support for options trading
 //!* WASM compilation allowing it to be embedded in static web pages
-
-//!**Warning**: the project is still in development and a lot of breaking changes are being made.
-
-
-//!### Community
-
+//!
+//! **Warning**: the project is still in development and a lot of breaking changes are being made.
+//!
+//! ### Community
+//!
 //!There is a project slack, [click here](https://join.slack.com/t/openlimitsworkspace/shared_invite/zt-ifhd5tv4-9QL73tCB~fDcRJzDFEuOrQ) to join.
-
+//!
 //!### Testing
-
+//!
 //!In order to run the tests you will have to provide environment variables for the sandbox API of the exchanges, 
 //! you can use environment variables or use a `.env` file.
-
+//!
 //!### Sponsorship
-
+//!
 //!We invite industry participants to join us in sponsoring a new high quality open source standard for crypto trading APIs. [
 //! Nash](https://nash.io) is dedicating a maintainer and initial rewards for external contributors that close issues. 
 //! Look for the wiki [rewards table](https://github.com/nash-io/openlimits/wiki/Rewards-sizes) and for reward size labels on open issues.
@@ -45,7 +44,7 @@ use crate::exchange::shared::Result;
 pub mod exchange;
 pub mod model;
 pub mod prelude;
-mod errors;
+pub mod errors;
 
 /// Can be used to initiate exchanges
 pub struct OpenLimits {}
@@ -55,4 +54,3 @@ impl OpenLimits {
         Ok(E::new(parameters).await?)
     }
 }
-
