+++ src/model/request/get_historic_trades_request.rs
@@ -1,7 +1,7 @@
 use derive_more::Constructor;
 use serde::Deserialize;
 use serde::Serialize;
-use super::Paginator;
+use crate::model::Paginator;
 
 /// This struct represents the historic of the trades
 #[derive(Serialize, Deserialize, Clone, Constructor, Debug)]
