+++ src/model/request/get_price_ticker_request.rs
