+++ src/model/request/order_book_response.rs
@@ -1,7 +1,7 @@
 use derive_more::Constructor;
 use serde::Deserialize;
 use serde::Serialize;
-use super::AskBid;
+use crate::model::AskBid;
 
 /// This struct represents an order book response
 #[derive(Serialize, Deserialize, Clone, Constructor, Debug, Default, PartialEq)]
